refactor:CatAnimationのアクションのアニメーションを分離、型定義
